don't spin on container kill

+++ agent.go
@@ -239,7 +239,7 @@ func (a *HostAgent) dockerTerminate(dockerID string) error {
 		return err
 	}
 
-	if err = dc.KillContainer(dockerID); err != nil {
+	if err = dc.KillContainer(dockerID); err != nil && !strings.Contains(err.Error(), "No such container") {
 		glog.Errorf("unable to kill container %s: %v", dockerID, err)
 		return err
 	}
